requirements.txt
aiofiles==0.8.0
anyio==3.5.0
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
docxcompose==1.3.4
docxtpl==0.15.2
fastapi==0.75.0
h11==0.13.0
httptools==0.4.0
idna==3.3
Jinja2==3.0.3
lxml==4.8.0
MarkupSafe==2.1.1
pydantic==1.9.0
python-docx==0.8.11
python-dotenv==0.19.2
python-multipart==0.0.5
PyYAML==6.0
requests==2.27.1
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
tqdm==4.63.0
typing-extensions==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8
websockets==10.2
